Fix Bug with radioIn nullptr
The ChannelAccessExtended/ChannelControlExtended modules were used to
comply with WiFi, but introduced a bug with registerHost.
With the "ordinary" ChannelControl setting the radioInGate parameter to
the gate "radioIn" of the host if the parameter was null, the extended
version did not do so.
A fix was introduced by also checking the radioInGate parameter for a
null pointer and possibly reassigning the radioInGate in the extended
version.